requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-login==0.6.2
Werkzeug==2.3.0

